+++ contracts/ecosystem/andromeda-mirror/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "andromeda-mirror-wrapped-cdp"
+name = "andromeda-mirror"
 version = "0.1.0"
 edition = "2018"
