+++ packages/andromeda-ecosystem/src/lib.rs
@@ -1,6 +1,6 @@
 pub mod anchor_earn;
 pub mod anchor_lend;
 pub mod astroport;
-pub mod mirror_wrapped_cdp;
+pub mod mirror;
 pub mod swapper;
 pub mod vault;
